separate non-zone related handlers
